requirements/base.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
click==8.2.1
Django==5.2.6
django-autoslug==1.9.9
django-cors-headers==4.9.0
django-elasticsearch-dsl==9.0
django-filter==25.1
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
elastic-transport==9.1.0
elasticsearch==9.1.1
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
